+++ lib/fsblockstore/longtail_fsblockstore.c
@@ -672,7 +672,7 @@ static int FSBlockStore_PutStoredBlock(
 
 static int FSBlockStore_PreflightGet(
     struct Longtail_BlockStoreAPI* block_store_api,
-    uint64_t chunk_count,
+    uint32_t chunk_count,
     const TLongtail_Hash* chunk_hashes)
 {
     MAKE_LOG_CONTEXT_FIELDS(ctx)
@@ -935,7 +935,7 @@ static int FSBlockStore_GetIndexSync(
 
 static int FSBlockStore_GetExistingContent(
     struct Longtail_BlockStoreAPI* block_store_api,
-    uint64_t chunk_count,
+    uint32_t chunk_count,
     const TLongtail_Hash* chunk_hashes,
     uint32_t min_block_usage_percent,
     struct Longtail_AsyncGetExistingContentAPI* async_complete_api)
@@ -966,7 +966,7 @@ static int FSBlockStore_GetExistingContent(
     struct Longtail_StoreIndex* existing_store_index;
     err = Longtail_GetExistingStoreIndex(
         store_index,
-        (uint32_t)chunk_count,
+        chunk_count,
         chunk_hashes,
         min_block_usage_percent,
         fsblockstore_api->m_DefaultMaxBlockSize,
